[BPF] split fib.h into co-re and legacy
Going forward, we will maintain co-re and will eventually deprecate and
remove the legacy one. That will get us back to a single fib.h.
